Solaris compilation fixes.
Result of the first attempt at compiling UFTPS under Solaris.  Basically, it was
just a matter of adding missing header files.

Some aid to the compilation process under different system has yet to be
implemented.

+++ log.c
@@ -44,7 +44,7 @@ static void message (      int      show_error,
 
         error_code = errno;
 
-        printf("(%5d) [%-7s] ", getpid(), severity);
+        printf("(%5d) [%-7s] ", (int) getpid(), severity);
         vprintf(msg, args);
 
         if (show_error && error_code != 0)
